Implementing Tokenizer and Parser for search, fixing the bug for search by price

+++ HumanitarianApp/app/src/main/java/com/example/ievent/database/data_manager/EventDataManager.java
@@ -95,7 +95,7 @@ public class EventDataManager {
      * @param listener the listener to handle the result
      */
     public synchronized void getGreaterThan(double price, EventDataListener listener) {
-        Query q = eventRef.whereGreaterThanOrEqualTo("price",price).limit(30);
+        Query q = eventRef.whereGreaterThanOrEqualTo("price",(double)(price)).limit(30);
         HandleQuery(q, listener);
     }
